refactor: split game region minimap into part files (#3878)
Extract painter and zoom/toggle chrome from the 522-line minimap widget
into library parts for Phase 3 flame modularization; behavior unchanged.
